organize files packages and dependencies
