+++ package.json
@@ -57,6 +57,7 @@
     "@oclif/plugin-help": "^2.2.3",
     "ajv": "^6.12.3",
     "awilix": "^4.2.6",
+    "axios": "^0.19.2",
     "bls-signatures": "^0.1.9",
     "chalk": "^4.0.0",
     "dash": "3.16.0",
@@ -75,6 +76,7 @@
     "pretty-bytes": "^5.3.0",
     "pretty-ms": "^7.0.0",
     "public-ip": "^4.0.1",
+    "qs": "^6.9.4",
     "rimraf": "^3.0.2",
     "rxjs": "latest",
     "semver": "^7.3.2",
